requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pdf417decoder==1.0.8
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pyzbar==0.1.9
six==1.17.0
zxing==1.0.3
